+++ program/src/syscalls/definitions.rs
@@ -11,7 +11,7 @@ pub use solana_define_syscall::definitions::{
     sol_curve_group_op, sol_curve_multiscalar_mul, sol_curve_pairing_map, sol_curve_validate_point,
     sol_get_clock_sysvar, sol_get_epoch_rewards_sysvar, sol_get_epoch_schedule_sysvar,
     sol_get_epoch_stake, sol_get_fees_sysvar, sol_get_last_restart_slot, sol_get_rent_sysvar,
-    sol_get_sysvar, sol_keccak256, sol_remaining_compute_units, BigModExpParams,
+    sol_get_sysvar, sol_keccak256, sol_remaining_compute_units,
 };
 #[cfg(target_feature = "static-syscalls")]
 pub use solana_define_syscall::sys_hash;
